requirements.txt
appdirs==1.4.3
Babel==2.4.0
boto==2.48.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
dj-database-url==0.4.2
Django==1.8
django-admin-sortable==2.1
django-ckeditor==4.5.0
django-grappelli==2.10.1
django-localflavor==1.5
django-phonenumber-field==1.3.0
django-storages==1.6.5
django-suit==0.2.25
future==0.16.0
geocoder==1.38.1
geopy==1.11.0
gunicorn==19.7.1
idna==2.6
olefile==0.44
packaging==16.8
phonenumberslite==8.5.0
Pillow==4.1.1
psycopg2==2.7.3.2
pyparsing==2.2.0
pytz==2017.2
ratelim==0.1.6
requests==2.18.4
six==1.10.0
ua-parser==0.8.0
urllib3==1.22
user-agents==1.1.0
